fixed the error of calculating the average speed improvement
Former-commit-id: 554299996213a24c8e9148344f195f5437a21c84

+++ boxFilter/test_boxFilter.cpp
@@ -129,7 +129,7 @@ time_boxFilter( carp::opencl::device & device, T0 & pool )
             carp::Timing::print( "cvtColor", elapsed_time_cpu, elapsed_time_gpu );
             
             if (elapsed_time_gpu > 1)
-                sum_quotient += elapsed_time_cpu / elapsed_time_gpu;
+                sum_quotient += static_cast<double>(elapsed_time_cpu) / elapsed_time_gpu;
 
         } // for size
